Split ViewDistributionItem and associated protocols/extensions into a separate file
