+++ Paralayout/UIView+Distribution.swift
@@ -16,15 +16,6 @@
 
 import UIKit
 
-/// A means of getting a `SubviewDistributionItem`: either a UIView, or a number as `.fixed` or `.flexible`.
-public protocol ViewDistributionSpecifying {
-
-    var distributionItem: ViewDistributionItem { get }
-
-}
-
-// MARK: -
-
 /// A direction for subview distribution.
 public enum ViewDistributionAxis {
 
@@ -34,9 +25,9 @@ public enum ViewDistributionAxis {
     /// The vertical direction, meaning views will be distributed top-to-bottom.
     case vertical
 
-    // MARK: - Private Methods
+    // MARK: - Internal Methods
 
-    fileprivate func amount(of insets: UIEdgeInsets) -> CGFloat {
+    internal func amount(of insets: UIEdgeInsets) -> CGFloat {
         switch self {
         case .horizontal:
             return insets.horizontalAmount
@@ -45,7 +36,7 @@ public enum ViewDistributionAxis {
         }
     }
 
-    fileprivate func size(of rect: CGRect) -> CGFloat {
+    internal func size(of rect: CGRect) -> CGFloat {
         switch self {
         case .horizontal:
             return rect.width
@@ -157,304 +148,9 @@ public enum ViewSpreadingBehavior {
 
 }
 
-/// An element of a horizontal or vertical distribution.
-public enum ViewDistributionItem: ViewDistributionSpecifying {
-
-    /// A UIView, with adjustments to how much space it should take up.
-    case view(UIView, UIEdgeInsets)
-
-    /// A constant spacer between two other elements.
-    case fixed(CGFloat)
-
-    /// Proportional spacer, a fraction of the space not taken up by UIViews or fixed spacers.
-    case flexible(CGFloat)
-
-    // MARK: - Public Static Methods
-
-    /// Filter invisible views (nil, uninstalled, hidden, or transparent) from a distribution, and collapse adjacent
-    /// spacers (preferring larger ones).
-    /// - parameter distribution: An array of optional distribution specifiers: either a UIView, or a number as `.fixed`
-    /// or `.flexible`.
-    /// - returns: An array of DistributionItems, without any invisible views, or sequential `.fixed` or `.flexible`
-    /// spacers.
-    public static func collapsing(_ distribution: [ViewDistributionSpecifying?]) -> [ViewDistributionItem] {
-        var collapsedItems = [ViewDistributionItem]()
-
-        func isViewVisible(_ view: UIView) -> Bool {
-            if view.superview == nil || view.isHidden || view.alpha == 0 {
-                return false
-
-            } else if let label = view as? UILabel {
-                if let text = label.text {
-                    return !text.isEmpty
-                } else {
-                    return false
-                }
-
-            } else if let imageView = view as? UIImageView {
-                return (imageView.image != nil)
-
-            } else {
-                return true
-            }
-        }
-
-        for distributionSpecifier in distribution {
-            // Filter out anything that's nil.
-            guard let item = distributionSpecifier?.distributionItem else {
-                continue
-            }
-
-            switch item {
-            case .view(let view, _):
-                // Filter out invisible views.
-                if isViewVisible(view) {
-                    collapsedItems.append(item)
-                }
-
-            case .fixed(let fixedSpace):
-                if let previousItem = collapsedItems.last {
-                    switch previousItem {
-                    case .view:
-                        // Space after a visible view: append.
-                        collapsedItems.append(item)
-
-                    case .fixed(let previousFixedSpace):
-                        // Fixed after fixed: replace if larger.
-                        if fixedSpace > previousFixedSpace {
-                            collapsedItems.removeLast()
-                            collapsedItems.append(item)
-                        }
-
-                    case .flexible:
-                        // Fixed after flexible: skip.
-                        break
-                    }
-
-                } else {
-                    // This is the first item: append.
-                    collapsedItems.append(item)
-                }
-
-            case .flexible(let flexibleSpace):
-                if let previousItem = collapsedItems.last {
-                    switch previousItem {
-                    case .view:
-                        // Space after a visible view: append.
-                        collapsedItems.append(item)
-
-                    case .fixed:
-                        // Flexible after fixed: replace.
-                        collapsedItems.removeLast()
-                        collapsedItems.append(item)
-
-                    case .flexible(let previousFlexibleSpace):
-                        // Flexible after flexible: replace if larger.
-                        if flexibleSpace > previousFlexibleSpace {
-                            collapsedItems.removeLast()
-                            collapsedItems.append(item)
-                        }
-                    }
-
-                } else {
-                    // This is the first item: append.
-                    collapsedItems.append(item)
-                }
-            }
-        }
-
-        // Trim out fixed space at the start or finish.
-        if let firstItem = collapsedItems.first {
-            switch firstItem {
-            case .fixed:
-                collapsedItems.removeFirst()
-            case .flexible, .view:
-                break
-            }
-        }
-
-        if let lastItem = collapsedItems.last {
-            switch lastItem {
-            case .fixed:
-                collapsedItems.removeLast()
-            case .flexible, .view:
-                break
-            }
-        }
-
-        // All set.
-        return collapsedItems
-    }
-
-    /// Filter invisible views (nil, uninstalled, hidden, or transparent) from a distribution, and collapse adjacent
-    /// spacers (preferring larger ones).
-    /// - parameter distribution: An series of optional distribution specifiers: either a UIView, or a number as
-    /// `.fixed` or `.flexible`.
-    /// - returns: An array of DistributionItems, without any invisible views, or sequential `.fixed` or `.flexible`
-    /// spacers.
-    public static func collapsing(_ distribution: ViewDistributionSpecifying? ...) -> [ViewDistributionItem] {
-        return collapsing(distribution)
-    }
-
-    // MARK: - Properties
-
-    /// Itself: `DistributionItem` trivially conforms to `ViewDistributionSpecifying`.
-    public var distributionItem: ViewDistributionItem {
-        return self
-    }
-
-    /// Whether or not this item is flexible.
-    public var isFlexible: Bool {
-        switch self {
-        case .view, .fixed:
-            return false
-        case .flexible:
-            return true
-        }
-    }
-
-    // MARK: - Private Methods
-
-    /// Maps the specifiers to their provided items, and adds implied flexible spacers as necessary.
-    /// If no spacers are included, equal flexible spacers are inserted between all views; if no `.flexible` spacers are
-    /// included, two equal ones are added to the beginning and end.
-    /// - returns: An array of DistributionItems suitable for layout and/or measurement, and tallies of all fixed and
-    /// flexible space. If the distribution is invalid (no views, any view not a subview of the superview, or any view
-    /// repeated in the distribution), returns an empty array.
-    fileprivate static func items(
-        impliedIn distribution: [ViewDistributionSpecifying],
-        axis: ViewDistributionAxis,
-        superview: UIView?
-    ) -> (items: [ViewDistributionItem], totalFixedSpace: CGFloat, flexibleSpaceDenominator: CGFloat) {
-        var distributionItems = [ViewDistributionItem]()
-        var totalViewSize: CGFloat = 0
-        var totalFixedSpace: CGFloat = 0
-        var totalFlexibleSpace: CGFloat = 0
-
-        var subviewsToDistribute = Set<UIView>()
-
-        // Map the specifiers to items, tallying up space along the way.
-        for specifier in distribution {
-            let item = specifier.distributionItem
-            let layoutSize = item.layoutSize(along: axis)
-
-            switch item {
-            case .view(let view, _):
-                // Validate the view.
-                guard superview == nil || view.superview === superview else {
-                    fatalError("\(view) is not a subview of \(String(describing: superview))!")
-                }
-
-                guard !subviewsToDistribute.contains(view) else {
-                    fatalError("\(view) is included twice in \(distribution)!")
-                }
-
-                subviewsToDistribute.insert(view)
-
-                totalViewSize += layoutSize
-
-            case .fixed:
-                totalFixedSpace += layoutSize
-
-            case .flexible:
-                totalFlexibleSpace += layoutSize
-            }
-
-            distributionItems.append(item)
-        }
-
-        // Exit early if no subviews were provided.
-        guard subviewsToDistribute.count > 0 else {
-            return ([], 0, 0)
-        }
-
-        // Insert flexible space if necessary.
-        if totalFlexibleSpace == 0 {
-            if totalFixedSpace == 0 {
-                // No spacers at all: insert `1.flexible` between all items.
-                for i in 0 ..< (distributionItems.count + 1) {
-                    distributionItems.insert(1.flexible, at: i * 2)
-                    totalFlexibleSpace += 1
-                }
-            } else {
-                // Only fixed spacers: add `1.flexible` on both ends.
-                distributionItems.insert(1.flexible, at: 0)
-                distributionItems.append(1.flexible)
-                totalFlexibleSpace += 2
-            }
-        }
-
-        return (distributionItems, totalFixedSpace + totalViewSize, totalFlexibleSpace)
-    }
-
-    /// Returns the length of the DistributionItem (`axis` and `multiplier` are relevant only for `.view` and
-    /// `.flexible` items, respectively).
-    fileprivate func layoutSize(along axis: ViewDistributionAxis, multiplier: CGFloat = 1) -> CGFloat {
-        switch self {
-        case .view(let view, let insets):
-            return axis.size(of: view.untransformedFrame) - axis.amount(of: insets)
-
-        case .fixed(let margin):
-            return margin
-
-        case .flexible(let space):
-            return space * multiplier
-        }
-    }
-
-}
-
 // MARK: -
 
-extension CGFloat {
-
-    /// Use the value as a fixed spacer in a distribution.
-    public var fixed: ViewDistributionItem {
-        return .fixed(self)
-    }
-
-    /// Use the value as a flexible (proportional) spacer in a distribution.
-    public var flexible: ViewDistributionItem {
-        return .flexible(self)
-    }
-
-}
-
-extension Double {
-
-    /// Use the value as a fixed spacer in a distribution.
-    public var fixed: ViewDistributionItem {
-        return .fixed(CGFloat(self))
-    }
-
-    /// Use the value as a flexible (proportional) spacer in a distribution.
-    public var flexible: ViewDistributionItem {
-        return .flexible(CGFloat(self))
-    }
-
-}
-
-extension Int {
-
-    /// Use the value as a fixed spacer in a distribution.
-    public var fixed: ViewDistributionItem {
-        return .fixed(CGFloat(self))
-    }
-
-    /// Use the value as a flexible (proportional) spacer in a distribution.
-    public var flexible: ViewDistributionItem {
-        return .flexible(CGFloat(self))
-    }
-
-}
-
-extension UIView : ViewDistributionSpecifying {
-
-    /// Adopt `ViewDistributionSpecifying`, making it possible to include UIView instances directly in distributions
-    /// passed to `apply[Vertical,Horizontal]Distribution()`.
-    public var distributionItem: ViewDistributionItem {
-        return .view(self, .zero)
-    }
+extension UIView {
 
     /// Arrange subviews according to a distribution with fixed and/or flexible spacers. Examples:
     ///
